requirements.txt
Django==5.1.3
python-dotenv==1.2.2
